Chore(tools): Initial project setup and build system
Sets up the initial project structure, cli tools and build system for
the Erika game engine project. Currently the Premake file declares a
workspace for getting it building, but in the future this will be
removed and it will be assumed the user imports the Erika project as a
subproject/module.